% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanPaginate.R
\name{can_paginate}
\alias{can_paginate}
\title{Paginator client}
\description{
Paginator client
}
\details{
**Methods**
  \describe{
    \item{`get(path, query, disk, stream, ...)`}{
      Make a GET request
    }
  }

Possible parameters (not all are allowed in each HTTP verb):
\itemize{
 \item path - URL path, appended to the base URL
 \item query - query terms, as a list
 \item body - body as an R list
 \item encode - one of form, multipart, json, or raw
}
}
\examples{
(x <- can_paginate(
  by = "query_params",
  limit_param = "limit",
  offset_param = "offset",
  limit = 5000))
}
