% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canpaginate-package.R
\docType{package}
\name{canpaginate-package}
\alias{canpaginate-package}
\alias{canpaginate}
\title{canpaginate}
\description{
Pagination Helper For Working With 'REST' 'APIs'
}
\section{Package API}{

x
}

\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}
}
\keyword{package}
